Finished Robot Code
Everything still needs testing, and I definitely made some mistakes with
the arm solenoids as I don't have the robot to look at, but it's done as
far as I know.  Added Util/JoystickAxisTrigger to allow the right
trigger on the gamepad (which is analog) to function as a button.